Relocating Telemetry
