+++ pkg/telemetry/telemetry_test.go
@@ -1,0 +1,10 @@
+package telemetry
+
+import (
+	"testing"
+)
+
+func NewTelemetry(t *testing.T) {
+	m := New()
+	m.Srv.Reset()
+}
